Python_exercises/Data-Science-Academy/requirements.txt
absl-py==2.1.0
asttokens==2.4.1
astunparse==1.6.3
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.2.0
cycler==0.12.1
decorator==5.1.1
executing==2.0.1
flatbuffers==24.3.7
fonttools==4.49.0
gast==0.4.0
google-auth==2.29.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.62.1
h5py==3.10.0
idna==3.6
ipython==8.22.2
jax==0.4.25
jedi==0.19.1
joblib==1.3.2
keras==2.12.0
kiwisolver==1.4.5
libclang==18.1.1
Markdown==3.6
MarkupSafe==2.1.5
matplotlib==3.8.3
matplotlib-inline==0.1.6
ml-dtypes==0.3.2
numpy==1.23.5
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
pandas==2.2.1
parso==0.8.3
patsy==0.5.6
pillow==10.2.0
prompt-toolkit==3.0.43
protobuf==4.25.3
pure-eval==0.2.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
Pygments==2.17.2
pyparsing==3.1.2
python-dateutil==2.8.2
pytz==2024.1
requests==2.31.0
requests-oauthlib==1.4.0
rsa==4.9
scikit-learn==1.4.1.post1
scipy==1.12.0
seaborn==0.13.2
six==1.16.0
stack-data==0.6.3
statsmodels==0.14.1
tensorboard==2.12.3
tensorboard-data-server==0.7.2
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.3.0
tornado==6.4
traitlets==5.14.1
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
wcwidth==0.2.13
Werkzeug==3.0.1
wrapt==1.14.1
